feat: optimize performance

+++ pedrodb/include/pedrodb/options.h
@@ -6,7 +6,7 @@
 
 namespace pedrodb {
 struct Options {
-  int16_t max_open_files = 16;
+  int8_t max_open_files = 16;
   size_t read_cache_bytes = kMaxFileBytes;
   size_t compaction_threshold_bytes = kMaxFileBytes / 2;
   Duration sync_interval = Duration::Seconds(10);
